🎨 로그아웃/회원탈퇴 버튼을 마이페이지로 이동시키기 위해 컴포넌트 생성
